Rename core app to game
