--- core/__init__.py
@@ -1,1 +1,0 @@
-"""Core app package."""
