--- core/apps.py
@@ -1,10 +1,0 @@
-"""Application configuration for the core app."""
-
-from django.apps import AppConfig
-
-
-class CoreConfig(AppConfig):
-    """Configuration for the core app."""
-
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "core"
