--- core/enums/__init__.py
@@ -1,6 +1,0 @@
-"""Enumeration classes for the core application."""
-
-from core.enums.country_enum import Country
-from core.enums.jp_prefecture_enum import JPPrefecture
-
-__all__ = ["Country", "JPPrefecture"]
