+++ game/__init__.py
@@ -1,0 +1,1 @@
+"""Game app package."""
