+++ game/enums/__init__.py
@@ -1,0 +1,6 @@
+"""Enumeration classes for the game application."""
+
+from game.enums.country_enum import Country
+from game.enums.jp_prefecture_enum import JPPrefecture
+
+__all__ = ["Country", "JPPrefecture"]
