+++ game/enums/jp_prefecture_enum.py
