+++ game/tests.py
@@ -1,3 +1,3 @@
-"""Tests for the core app."""
+"""Tests for the game app."""
 
 from django.test import TestCase  # noqa: F401
