+++ game/views.py
@@ -1,3 +1,3 @@
-"""Views for the core app."""
+"""Views for the game app."""
 
 from django.shortcuts import render  # noqa: F401
